feat: adjust sound paths
