+++ pages/index.page.tsx
@@ -4,8 +4,11 @@ import { useEffect } from 'react'
 import { useGlitch } from 'react-powerglitch'
 import Sound from 'react-sound'
 import Equalizer from '~/src/components/Equalizer'
-import { SoundClickButton, SoundGlitch } from '~/src/utils/Sounds'
-import { settingsSound } from '~/src/utils/settingsHome'
+import {
+  SoundClickButton,
+  SoundGlitch,
+  settingsSound,
+} from '~/src/utils/sounds'
 
 import * as S from './styles'
