+++ src/layout/components/NavBar/Desktop/index.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 import Equalizer from '~/src/components/Equalizer'
-import { SoundClickButton } from '~/src/utils/Sounds'
+import { SoundClickButton } from '~/src/utils/sounds'
 
 import { navBarMock } from '../Navbar.mock'
 import * as S from './styles'
