--- src/utils/settingsHome.ts
@@ -1,5 +1,0 @@
-export const settingsSound = {
-  url: 'https://cdn.discordapp.com/attachments/778024116140769331/1157484834583760997/Techno-Mania-Glitch-Effect-Sound.MP3?ex=6518c741&is=651775c1&hm=203d6e258d07d03650ee4ca4ea977e2825e6d43c833fc5d8e79f41b9704ee6c7&',
-  playbackRate: 2,
-  volume: 10,
-}
